Added mouse events to NodeGraphGraphicsView
NOTE: I think we should be using the NGGraphicsView to check for other
nodes already present in the scene, but I'm not sure.

Ultimately, we'll want the behaviour of: you can create a new node with
the mouse, but if you click on an existing node, that one is selected
for editing.